# backend/docs/openapi-gpt-action.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Isekai Lucas Engine GPT Actions",
    "version": "0.22.0",
    "description": "Schema normal para GPT Actions. Expone endpoints de lectura, busqueda, preview y mutacion central controlada. No expone rollback, admin ni mutadores peligrosos de combate/misiones."
  },
  "servers": [
    {
      "url": "https://isekai-lucas-engine.onrender.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "GenericResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "additionalProperties": true
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/context/full": {
      "get": {
        "operationId": "getFullContext",
        "summary": "Obtiene contexto vivo amplio con logs resumidos por defecto.",
        "responses": {
          "200": {
            "description": "Contexto vivo cercano.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        },
        "description": "Lectura amplia para debug. Por defecto limita recentEventLogs a 10 y no incluye mechanicalChanges para evitar ResponseTooLargeError. Para debug pesado usar includeMechanicalChanges=true&logLimit=30.",
        "parameters": [
          {
            "name": "gameId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "isekai_lucas_main"
            }
          },
          {
            "name": "logLimit",
            "in": "query",
            "required": false,
            "description": "Cantidad de logs recientes. Default 10, maximo 30.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 30,
              "default": 10
            }
          },
          {
            "name": "includeMechanicalChanges",
            "in": "query",
            "required": false,
            "description": "Incluir mechanicalChanges completos en recentEventLogs. Default false para evitar respuestas demasiado grandes.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ]
      }
    },
...
    "/api/context/compact": {
      "get": {
        "operationId": "getCompactContext",
        "summary": "Obtiene contexto compacto para narrar el turno sin exceder limites de GPT Actions.",
        "description": "Lectura principal del narrador: estado vivo, ubicacion, Lucas, NPCs cercanos, misiones, clima, trabajo, combate activo y alertas de coherencia.",
        "parameters": [
          {
            "name": "gameId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "isekai_lucas_main"
            }
          },
          {
            "name": "npcLimit",
            "in": "query",
            "required": false,
            "description": "Maximo de NPCs cercanos.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 30
            }
          },
          {
            "name": "memoryLimit",
            "in": "query",
            "required": false,
            "description": "Maximo de memorias relevantes.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 12
            }
          },
          {
            "name": "rumorLimit",
            "in": "query",
            "required": false,
            "description": "Maximo de rumores activos.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 10
            }
          },
          {
            "name": "missionLimit",
            "in": "query",
            "required": false,
            "description": "Maximo de misiones activas/disponibles.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 15
            }
          },
          {
            "name": "eventLimit",
            "in": "query",
            "required": false,
            "description": "Maximo de eventos vivos.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 12
            }
          },
          {
            "name": "logLimit",
            "in": "query",
            "required": false,
            "description": "Maximo de logs resumidos.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contexto compacto listo para GPT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}